requirements.txt
Django==3.1.6
djangorestframework==3.12.2
Pillow==8.1.0
djoser==2.1.0
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework-simplejwt==4.6.0




